Removed ServerModel as it was too general. Replaced it with Agent and Domain to handle specific interactions with their routes
